Move useKeyboardArrows in its own file
