+++ src/hooks/useKeyboardArrows.js
@@ -1,0 +1,26 @@
+import { useEffect } from 'react'
+import keycodes from '../keycodes'
+
+export default function useArrows({ onUp, onLeft, onDown, onRight } = {}) {
+  useEffect(() => {
+    const actions = [
+      [keycodes.up, onUp],
+      [keycodes.left, onLeft],
+      [keycodes.down, onDown],
+      [keycodes.right, onRight],
+    ]
+    const onKeyDown = e => {
+      for (const [keyCode, cb] of actions) {
+        if (cb && e.keyCode === keyCode) {
+          e.preventDefault()
+          cb()
+          break
+        }
+      }
+    }
+    document.addEventListener('keydown', onKeyDown)
+    return () => {
+      document.removeEventListener('keydown', onKeyDown)
+    }
+  }, [onUp, onLeft, onDown, onRight])
+}
